Move Home and Quote component to the component directory
