+++ src/components/Quote.js
